requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
pymongo==4.3.3
motor==3.1.1
python-dotenv==1.0.0
jinja2==3.1.2
aiosmtplib==2.0.1
pydantic==1.10.7
email-validator==2.0.0
